refactor(memory_cluster): 重构模块结构
