+++ src/memory.rs
@@ -1,6 +1,6 @@
 pub mod algo;
+pub mod cluster;
 pub mod embedding;
-pub mod memory_cluster;
 pub mod memory_links;
 pub mod memory_note;
 pub mod query;
